requirements.txt
asgiref==3.6.0
gunicorn==20.1.0
sqlparse==0.4.4
whitenoise==6.4.0
Django==5.1
django_crispy_forms==2.3
django_tables2==2.7.0
unfold==1.2.0
